pyproject.toml
[tool.poetry]
name = "coursework-6-skymarket"
version = "0.1.0"
description = ""
authors = ["TatianaBelova333 <101121235+TatianaBelova333@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "coursework_6_skymarket"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
psycopg2 = "^2.9.3"
drf-spectacular = "^0.24.0"
django-cors-headers = "^3.13.0"
python-dotenv = "^0.21.0"
django-phonenumber-field = "^7.0.0"
phonenumbers = "^8.12.55"
Pillow = "^9.2.0"
djoser = "^2.1.0"
django-filter = "^22.1"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
